removes compatibility with cross domain localstorage and changes to sessionScope format to match auth-react
